backup_requirements.txt
asgiref==3.6.0
awsebcli==3.20.5
botocore==1.29.81
cement==2.8.2
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.3
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
filelock==3.9.0
idna==3.4
jmespath==1.0.1
npm==0.1.1
pathspec==0.10.1
platformdirs==3.0.0
pypiwin32==223
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
termcolor==1.1.0
tzdata==2022.7
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.1.9
